feat: add `[keepMountedShowNext]` to Modal
