+++ src/components/Modal/index.ts
@@ -1,2 +1,2 @@
 export { default } from './Modal';
-export * from './Modal.type';
+export type { ModalProps } from './Modal';
